Added initial implementation of byteFieldOffset (issue #13) and did some refactoring.
Adding byteFieldOffset function requires a new PrimFields type family to be associated
with the PrimBytes type class. However, GHC does not support type family instances in
.hs-boot files, so my Backend instances have broken. To workaround this, I have split
Backend module into two. A new module BackendI exports only the generic ToInstance instances
and does not need to be imported as .hs-boot. In the end, this has simplified the appearance
of the Backend module, because I don't need to write .hs-boot-friendly instance stubs.

Tests for PrimBytes are not here yet.